fix(orchestrator): fix module not found errors
Change the command to run the orchestrator.
The previous command led to uncanny and undesireable module not found errors.
The new command should be the right way to run a package, but be sure to define a main function inside the __init__.py file. The main function is the entrypoint of the package. The entrypoint's name can be changed in the [project.scripts] table inside pyproject.toml